# bl_lookup/templates/openapi.yml
openapi: 3.0.2
info:
  title: Biolink Model Lookup
  description: 'The <a href="https://biolink.github.io/biolink-model/">Biolink Model</a> defines a set of common concepts for use in Translator. These include semantic types for entities, as well as the relations between them. These concepts are organized into an inheritance hierarchy capturing different granularities of description. Furthermore, each concept contains metadata relating the concept to ontologies.

The Biolink Lookup Service provides a computational interface to the model, including access to previous versions. When the service is deployed, it queries the Biolink Github repository, and updates itself to access the latest version.'
  version: '1.0.0'
  contact:
    email: bizon@renci.org
    name: Chris Bizon
    x-id: https://github.com/cbizon
    x-role: responsible developer
  termsOfService: http://robokop.renci.org:7055/tos?service_long=Biolink+Model+Lookup&provider_long=the+Translator+Consortium
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  x-translator:
    component: Utility
    team:
      - Standards Reference Implementation Team
servers:
  - description: Default server
    url: https://bl-lookup-sri.renci.org/
tags:
  - name: translator
  - name: lookup
  - name: meta
paths:
  /bl/{concept}/ancestors:
    get:
      tags:
        - lookup
      summary: 'Return all ancestors (superclasses or superproperties) of concept, based on the BioLink Model YAML.'
      description: ''
      parameters:
        - in: path
          required: true
          name: concept
          default: biological_process
        - in: query
          name: version
          default: ~default version~
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /bl/{concept}/descendants:
    get:
      tags:
        - lookup
      summary: 'Return descendants (subclasses or subproperties) of concept, based on the BioLink Model YAML'
      description: ''
      parameters:
        - in: path
          required: true
          name: concept
          default: disease_or_phenotypic_feature
        - in: query
          name: version
          default: ~default version~
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /bl/{concept}/lineage:
    get:
      tags:
        - lookup
      summary: 'Return the ancestors and descendants of concept, based on the BioLink Model YAML'
      description: ''
      parameters:
        - in: path
          required: true
          name: concept
          default: biological_process
        - in: query
          name: version
          default: ~default version~
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /uri_lookup/{uri}:
    get:
      tags:
        - lookup
      summary: 'Given the URI for a property, return the corresponding BioLink Model slot. Checks the Biolink Model YAML for both slot_uri and mappings.'
      description: ''
      parameters:
        - in: path
          required: true
          name: uri
          example: RO:0002606
        - in: query
          name: version
          default: ~default version~
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /bl/{concept}:
    get:
      tags:
        - lookup
      summary: 'Get all properties for concept from the BioLink Model YAML.'
      description: ''
      parameters:
        - in: path
          required: true
          name: concept
          default: biological_process
        - in: query
          name: version
          default: ~default version~
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object

  /resolve_predicate:
    get:
      tags:
        - lookup
      summary: 'Resolve predicate CURIE(s) into Biolink model compliant counterpart(s), first checking the BioLink model for
      exact matches of a CURIE.  If the CURIE is from the Relation Ontology, the service will check its superproperties until
       an exact match is found.'
      description: 'Get Biolink model compliant predicate(s) using the passed predicate(s).'
      parameters:
        - in: query
          name: predicate
          schema:
            type: array
            items:
              type: string
            example:
              - SEMMEDDB:CAUSES
              - RO:0000052
        - in: query
          name: version
          schema:
            type: string
            example: "~default version~"
      responses:
        '200':
          description: Results
          content:
            application/json:
              schema:
                type: object
                example:
                  "SEMMEDDB:CAUSES":
                    "identifier": "RO:0002410"
                    "label": "causes"
                  "RO:0000052":
                    "identifier": "SEMMEDDB:AFFECTS"
                    "label": "affects"


  /versions:
    get:
      tags:
        - meta
      summary: 'Get available BioLink Model versions. These values can be used to parameterize other function calls.'
      description: ''
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string